requirements.txt
requests
pyyaml
tqdm
beautifulsoup4
